Correct some typing problem

+++ PokemonGo/RocketAPI/Window/Settings.cs
@@ -54,8 +54,8 @@ namespace PokemonGo.RocketAPI.Window
         public string pokeballMode => GetSetting() != string.Empty ? GetSetting() : "IV";
         public int cpToUseSuperBall => GetSetting() != string.Empty ? int.Parse(GetSetting()) : 250;
         public int cpToUseHyperBall => GetSetting() != string.Empty ? int.Parse(GetSetting()) : 500;
-        public int IVToUseSuperBall => GetSetting() != string.Empty ? int.Parse(GetSetting()) : 75;
-        public int IVToUseHyperBall => GetSetting() != string.Empty ? int.Parse(GetSetting()) : 90;
+        public float IVToUseSuperBall => GetSetting() != string.Empty ? float.Parse(GetSetting()) : 75;
+        public float IVToUseHyperBall => GetSetting() != string.Empty ? float.Parse(GetSetting()) : 90;
         public int travelMode => GetSetting() != string.Empty ? int.Parse(GetSetting()) : 1;
         public int deplacementsMaxDist => GetSetting() != string.Empty ? int.Parse(GetSetting()) : 5;
         public int maxPokestop => GetSetting() != string.Empty ? int.Parse(GetSetting()) : 300;
